separated the page style into a separate file
