+++ src/engraving/CMakeLists.txt
@@ -88,6 +88,8 @@ set(MODULE_SRC
     ${CMAKE_CURRENT_LIST_DIR}/style/styledef.h
     ${CMAKE_CURRENT_LIST_DIR}/style/textstyle.cpp
     ${CMAKE_CURRENT_LIST_DIR}/style/textstyle.h
+    ${CMAKE_CURRENT_LIST_DIR}/style/pagestyle.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/style/pagestyle.h
     ${CMAKE_CURRENT_LIST_DIR}/style/style.cpp
     ${CMAKE_CURRENT_LIST_DIR}/style/style.h
     )
